--- src/crrp-notification-worker/README.md
@@ -1,13 +1,0 @@
-# CRRP Notification worker
-
-This Application subscribe to CRRP Output Notification and update FIGARO databse
-
-### Reference Documentation
-For further reference, please consider the following sections:
-
-* [Official Apache Maven documentation](https://maven.apache.org/guides/index.html)
-* [Spring Boot Maven Plugin Reference Guide](https://docs.spring.io/spring-boot/docs/2.2.4.RELEASE/maven-plugin/)
-* [Spring Boot Actuator](https://docs.spring.io/spring-boot/docs/2.2.4.RELEASE/reference/htmlsingle/#production-ready)
-
-
-
