--- src/crrp-notification-worker/src/main/java/ca/bc/gov/open/pssg/rsbc/crrp/notification/worker/CrrpNotificationWorkerApplication.java
@@ -1,13 +1,0 @@
-package ca.bc.gov.open.pssg.rsbc.crrp.notification.worker;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class CrrpNotificationWorkerApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(CrrpNotificationWorkerApplication.class, args);
-	}
-
-}
