--- src/crrp-notification-worker/src/main/java/ca/bc/gov/open/pssg/rsbc/crrp/notification/worker/document/DocumentService.java
@@ -1,12 +1,0 @@
-package ca.bc.gov.open.pssg.rsbc.crrp.notification.worker.document;
-
-/**
- * Collection of services for documents.
- *
- * @author carolcarpenterjustice
- */
-public interface DocumentService {
-
-    DpsDataIntoFigaroResponse dpsDataIntoFigaro(DpsDataIntoFigaroRequestBody request);
-
-}
