--- src/crrp-notification-worker/src/main/resources/application.properties
@@ -1,5 +1,0 @@
-server.servlet.context-path=/crrpnotificationworker
-spring.rabbitmq.host: ${RABBITMQ_HOST:localhost}
-spring.rabbitmq.port: ${RABBITMQ_PORT:5672}
-spring.rabbitmq.username: ${RABBITMQ_USERNAME:guest}
-spring.rabbitmq.password: ${RABBITMQ_PASSWORD:guest}
