+++ src/spd-notification-worker/src/main/java/ca/bc/gov/open/pssg/rsbc/spd/notification/worker/Keys.java
@@ -1,0 +1,24 @@
+package ca.bc.gov.open.pssg.rsbc.spd.notification.worker;
+
+public class Keys {
+
+    private Keys() {}
+
+
+    /**
+     * The CRRP value for usage across the app
+     */
+    public static final String CRRP_VALUE = "CRRP";
+
+    /**
+     * DO NOT CHANGE - The name of the CRRP queue
+     */
+    public static final String CRRP_QUEUE_NAME = CRRP_VALUE + "_queue";
+
+    /**
+     * DO NOT CHANGE - The default output notification value.
+     */
+    public static final String OUTPUT_NOTIFICATION_VALUE = "outputNotification";
+
+
+}
