+++ src/spd-notification-worker/src/main/java/ca/bc/gov/open/pssg/rsbc/spd/notification/worker/SpdNotificationWorkerApplication.java
@@ -1,0 +1,13 @@
+package ca.bc.gov.open.pssg.rsbc.spd.notification.worker;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class SpdNotificationWorkerApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(SpdNotificationWorkerApplication.class, args);
+	}
+
+}
