+++ src/spd-notification-worker/src/main/java/ca/bc/gov/open/pssg/rsbc/spd/notification/worker/document/DocumentService.java
@@ -1,0 +1,12 @@
+package ca.bc.gov.open.pssg.rsbc.spd.notification.worker.document;
+
+/**
+ * Collection of services for documents.
+ *
+ * @author carolcarpenterjustice
+ */
+public interface DocumentService {
+
+    DpsDataIntoFigaroResponse dpsDataIntoFigaro(DpsDataIntoFigaroRequestBody request);
+
+}
