+++ src/spd-notification-worker/src/main/resources/application.properties
@@ -1,0 +1,5 @@
+server.servlet.context-path=/spdnotificationworker
+spring.rabbitmq.host: ${RABBITMQ_HOST:localhost}
+spring.rabbitmq.port: ${RABBITMQ_PORT:5672}
+spring.rabbitmq.username: ${RABBITMQ_USERNAME:guest}
+spring.rabbitmq.password: ${RABBITMQ_PASSWORD:guest}
